libia2/memory_maps: split memory maps out from permissive mode
This splits memory maps (`include/ia2_memory_maps.h`, `memory_maps.h`, `memory_maps.c`)
out of permissive mode (`include/permissive_mode.h`, `include/ia2_threads.h`, `threads.c`).

Previously, `permissive_mode.h` contained `log_memory_maps`,
a local function that printed the labeled memory maps to the log file.
Now this is moved to the public IA2 function `ia2_log_memory_maps` in `ia2_memory_maps.h`,
which can log to any `FILE*`, including `stderr` in a debugger.

Now that the whole public function is limited to just `ia2_log_memory_maps`,
all of the details previously exposed in `ia2_threads.h`
can be made private within `libia2` again, moved to `memory_maps.{c,h}`.

This pretty much only moves code around.  Other fixes, like handling locking
and errors when potentially called in a debugger better, are being left to later.

Fixes #568.
