+++ runtime/libia2/include/ia2_memory_maps.h
@@ -1,0 +1,5 @@
+#pragma once
+
+#include <stdio.h>
+
+void ia2_log_memory_maps(FILE* log);
